+++ VoxTerGenEngine/include/graphics/Texture2D.hpp
@@ -26,11 +26,29 @@ namespace TextureUtils
         GLenum target_;
 
     public:
-        Texture2D(std::string_view path, bool sRGB, bool generate_mipmaps,
-            GLenum wrap_s, GLenum wrap_t, GLenum min_filter, GLenum mag_filter);
+        Texture2D(
+            std::string_view path, 
+            bool sRGB, 
+            bool generate_mipmaps, 
+            bool flip_vertically, 
+            GLenum wrap_s, 
+            GLenum wrap_t, 
+            GLenum min_filter, 
+            GLenum mag_filter);
 
-        Texture2D(std::string_view cubemap_path, bool sRGB, bool generate_mipmaps,
-           GLenum wrap_s, GLenum wrap_t, GLenum wrap_r, GLenum min_filter, GLenum mag_filter);
+        Texture2D(
+            std::string_view path, 
+            GLsizei columns_n, 
+            GLsizei rows_n, 
+            const std::array<int, 6>& z_offsets, 
+            bool sRGB, 
+            bool generate_mipmaps, 
+            bool flip_vertically, 
+            GLenum wrap_s, 
+            GLenum wrap_t, 
+            GLenum wrap_r, 
+            GLenum min_filter, 
+            GLenum mag_filter);
 
         Texture2D(const Texture2D& other) = delete;
         Texture2D& operator=(const Texture2D& other) = delete;