+++ VoxTerGenEngine/include/render/ChunkMeshRenderPass.hpp
@@ -3,9 +3,13 @@
 
 #include "core/ResourceManager.hpp"
 
+#include "physics/AABB.hpp"
+
 #include "render/MeshRenderer.hpp"
+#include "render/MeshRenderData.hpp"
 
 #include "world/Chunk.hpp"
+#include "world/ChunkEvents.hpp"
 
 #include <glm/mat4x4.hpp>
 
@@ -20,11 +24,17 @@
 template<class... Ts>
 struct overloaded : Ts... { using Ts::operator()...; };
 
+struct ChunkData
+{
+	MeshRenderData mesh_render_data_;
+	AABB aabb_;
+};
+
 class ChunkMeshRenderPass
 {
 private:
 	const MeshRenderer& mesh_renderer_;
-	std::unordered_map<ChunkID, RenderData> chunks_render_data_;
+	std::unordered_map<ChunkID, MeshRenderData> chunks_data_;
 
 public:
 	ChunkMeshRenderPass(const MeshRenderer& mesh_renderer);