+++ VoxTerGenEngine/include/render/WorldRenderer.hpp
@@ -3,6 +3,9 @@
 
 #include "render/ChunkMeshRenderPass.hpp"
 #include "render/BlockHighlightRenderPass.hpp"
+#include "render/SkyboxRenderPass.hpp"
+
+#include "world/ChunkEvents.hpp"
 
 #include <glm/vec2.hpp>
 #include <glm/mat4x4.hpp>
@@ -12,6 +15,7 @@
 #include <queue>
 #include <optional>
 
+class Camera;
 class ResourceManager;
 
 struct RaycastResult;
@@ -22,6 +26,7 @@ private:
 	MeshRenderer mesh_renderer_;
 	ChunkMeshRenderPass chunk_mesh_render_pass_;
 	BlockHighlightRenderPass block_highlight_render_pass_;
+	SkyboxRenderPass skybox_render_pass_;
 
 public:
 	WorldRenderer();