+++ VoxTerGenEngine/include/world/ChunkManager.hpp
@@ -6,6 +6,7 @@
 #include "graphics/Camera.hpp"
 
 #include "world/Chunk.hpp"
+#include "world/ChunkEvents.hpp"
 #include "world/Block.hpp"
 
 #include <glm/vec2.hpp>
@@ -38,8 +39,6 @@ public:
 	glm::ivec3 AbsoluteBlockPos(glm::dvec3 position, glm::dvec3 pos_offset = { 0.0, 0.0, 0.0 }) const noexcept;
 
 	glm::ivec3 RelativeBlockPos(glm::dvec3 position, glm::dvec3 pos_offset = { 0.0, 0.0, 0.0 }) const noexcept;
-
-	void PushChunkIntoQueue(Chunk* chunk);
 
 	glm::ivec2 GetChunkCoords(glm::dvec3 pos) const noexcept;