+++ VoxTerGenEngine/src/graphics/Texture2D.cpp
@@ -12,6 +12,8 @@
 #include <algorithm>
 #include <cmath>
 #include <string_view>
+#include <cstring>
+#include <array>
 
 namespace TextureUtils
 {
@@ -41,13 +43,23 @@ namespace TextureUtils
         }
     }
 
-    Texture2D::Texture2D(std::string_view path, bool sRGB, bool generate_mipmaps, 
-        GLenum wrap_s, GLenum wrap_t, GLenum min_filter, GLenum mag_filter) 
-        : type_(TextureType::Texture2D), target_(GL_TEXTURE_2D)
+    Texture2D::Texture2D(
+        std::string_view path, 
+        bool sRGB, 
+        bool generate_mipmaps, 
+        bool flip_vertically, 
+        GLenum wrap_s, 
+        GLenum wrap_t, 
+        GLenum min_filter, 
+        GLenum mag_filter) 
+        : 
+        type_(TextureType::Texture2D), 
+        target_(GL_TEXTURE_2D)
     {
         glCreateTextures(GL_TEXTURE_2D, 1, &texture_id_);
         int n_components = 0;
         
+        stbi_set_flip_vertically_on_load(flip_vertically);
         std::unique_ptr<stbi_uc, decltype(&stbi_image_free)> data = { stbi_load(path.data(), &width_, &height_, &n_components, 0), stbi_image_free };
 
         if (!data.get())
@@ -88,19 +100,115 @@ namespace TextureUtils
         glTextureParameteri(texture_id_, GL_TEXTURE_MAG_FILTER, mag_filter);
     }
 
-    Texture2D::Texture2D(std::string_view cubemap_path, bool sRGB, bool generate_mipmaps,
-       GLenum wrap_s, GLenum wrap_t, GLenum wrap_r, GLenum min_filter, GLenum mag_filter)
-       : type_(TextureType::Cubemap), target_(GL_TEXTURE_CUBE_MAP)
+    Texture2D::Texture2D(
+        std::string_view path,
+        GLsizei columns_n, 
+        GLsizei rows_n, 
+        const std::array<int, 6>& z_offsets, 
+        bool sRGB, 
+        bool generate_mipmaps, 
+        bool flip_vertically, 
+        GLenum wrap_s, 
+        GLenum wrap_t, 
+        GLenum wrap_r, 
+        GLenum min_filter, 
+        GLenum mag_filter) 
+        : 
+        type_(TextureType::Cubemap), 
+        target_(GL_TEXTURE_CUBE_MAP)
     {
-       glCreateTextures(GL_TEXTURE_CUBE_MAP, 1, &texture_id_);
+        assert(columns_n != 0 && rows_n != 0);
 
-       // TODO
+        if (columns_n == 0 || rows_n == 0)
+        {
+            Logger::Log(LogLevel::INFO, "Invalid dimensions for cubemap '{}': columns_n {}, rows_n {}", path, columns_n, rows_n);
+            throw std::runtime_error("Invalid dimensions of cubemap");
+        }
 
-       glTextureParameteri(texture_id_, GL_TEXTURE_WRAP_S, wrap_s);
-       glTextureParameteri(texture_id_, GL_TEXTURE_WRAP_T, wrap_t);
-       glTextureParameteri(texture_id_, GL_TEXTURE_WRAP_R, wrap_r);
-       glTextureParameteri(texture_id_, GL_TEXTURE_MIN_FILTER, min_filter);
-       glTextureParameteri(texture_id_, GL_TEXTURE_MAG_FILTER, mag_filter);
+        glCreateTextures(GL_TEXTURE_CUBE_MAP, 1, &texture_id_);
+        int n_components = 0;
+        
+        stbi_set_flip_vertically_on_load(flip_vertically);
+        std::unique_ptr<stbi_uc, decltype(&stbi_image_free)> data = { stbi_load(path.data(), &width_, &height_, &n_components, 0), stbi_image_free };
+
+        if (!data.get())
+        {
+            glDeleteTextures(1, &texture_id_);
+            Logger::Log(LogLevel::INFO, "Failed to open cubemap file: {}", path.data());
+            throw std::runtime_error("Failed to open cubemap file: " + std::string(path) + " Error: " + stbi_failure_reason());
+        }
+
+        if (width_ <= 0 || height_ <= 0)
+        {
+            glDeleteTextures(1, &texture_id_);
+            Logger::Log(LogLevel::INFO, "Invalid dimensions for cubemap '{}': width {}, height {}", path, width_, height_);
+            throw std::runtime_error("Invalid dimensions of cubemap");
+        }
+        
+        GetTextureFormats(n_components, sRGB, &internal_format_, &data_format_);
+
+        const GLsizei face_width = width_ / columns_n;
+        const GLsizei face_height = height_ / rows_n;
+
+        assert(face_width == face_height);
+        assert(width_ % columns_n == 0);
+        assert(height_ % rows_n == 0);
+        assert(columns_n * rows_n == 6);
+
+        if (face_width != face_height || width_ % columns_n != 0 || height_ % rows_n != 0 || columns_n * rows_n != 6)
+        {
+            glDeleteTextures(1, &texture_id_);
+            Logger::Log(LogLevel::INFO, 
+                "Invalid data for cubemap '{}': width {}, height {}, face_width {}, face_height {}, columns_n {}, rows_n {} ", 
+                path, 
+                width_, 
+                height_, 
+                face_width, 
+                face_height, 
+                columns_n, 
+                rows_n);
+            throw std::runtime_error("Invalid data of cubemap");
+        }
+        
+        const GLsizei face_size = face_width;
+        const int levels = generate_mipmaps ? 1 + static_cast<int>(std::floor(std::log2(face_size))) : 1;
+        
+        GLint previous_unpack_alignment = 0;
+        glGetIntegerv(GL_UNPACK_ALIGNMENT, &previous_unpack_alignment);
+        glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
+
+        const std::size_t bytes_per_pixel = static_cast<std::size_t>(n_components);
+        const std::size_t face_width_bytes = static_cast<std::size_t>(face_size) * bytes_per_pixel;
+        const std::size_t atlas_width_bytes = face_width_bytes * static_cast<std::size_t>(columns_n);
+        const std::size_t buffer_size = static_cast<std::size_t>(face_size) * static_cast<std::size_t>(face_size) * bytes_per_pixel;
+        std::unique_ptr<stbi_uc[]> buffer = std::make_unique<stbi_uc[]>(buffer_size);
+
+        glTextureStorage2D(texture_id_, levels, internal_format_, face_size, face_size);
+
+        for (std::size_t i = 0; i < 6; ++i)
+        {
+            const std::size_t start_offset = (i / columns_n * static_cast<std::size_t>(face_size) * atlas_width_bytes) + ((i % columns_n) * face_width_bytes);
+
+            for (std::size_t row_n = 0; row_n < static_cast<std::size_t>(face_size); ++row_n)
+            {
+                std::memcpy(buffer.get() + (row_n * face_width_bytes), data.get() + start_offset + (row_n * atlas_width_bytes), face_width_bytes);
+            }
+
+            glTextureSubImage3D(texture_id_, 0, 0, 0, z_offsets[i], face_size, face_size, 1, data_format_, GL_UNSIGNED_BYTE, buffer.get());
+        }
+
+        if (generate_mipmaps)
+        {
+            glGenerateTextureMipmap(texture_id_);
+        }
+        
+        glPixelStorei(GL_UNPACK_ALIGNMENT, previous_unpack_alignment);
+        
+        glTextureParameteri(texture_id_, GL_TEXTURE_WRAP_S, wrap_s);
+        glTextureParameteri(texture_id_, GL_TEXTURE_WRAP_T, wrap_t);
+        glTextureParameteri(texture_id_, GL_TEXTURE_WRAP_R, wrap_r);
+        glTextureParameteri(texture_id_, GL_TEXTURE_MIN_FILTER, min_filter);
+        glTextureParameteri(texture_id_, GL_TEXTURE_MAG_FILTER, mag_filter);
     }
 
     Texture2D::Texture2D(Texture2D&& other) noexcept